Modified 'multiple args in subworkflow' inspection to include execution sections, added test.
