+++ src/main/resources/SnakemakeBundle.properties
@@ -25,15 +25,17 @@ INSP.NAME.shadow.multiple.settings=Only one argument after shadow section is per
 INSP.NAME.section.redeclaration=Section declaration above overrides this declaration
 INSP.NAME.section.redeclaration.message=Declaration of section ''{0}'' above overrides this declaration.
 
-# SmkSubworkflowMultipleArgsInspection
-INSP.NAME.subworkflow.multiple.args=Only one argument is allowed for subworkflow sections
+# SmkSectionMultipleArgsInspection
+INSP.NAME.section.multiple.args=Only one argument is allowed for this section.
+INSP.NAME.section.multiple.args.message=Only one argument is allowed for ''{0}'' section.
 
 # SmkSubworkflowRedeclarationInspection
 INSP.NAME.subworkflow.redeclaration=Only last subworkflow with the same name will be executed
 
 # SmkYetUndefinedNameInspection
 INSP.NAME.rule.or.checkpoint.name.yet.undefined=Rule or Checkpoint name has not been defined yet
-INSP.NAME.rule.or.checkpoint.name.yet.undefined.msg=Rule or Checkpoint name ''{0}'' has not been defined yet
+INSP.NAME.rule.or.checkpoint.name.yet.undefined.msg=Rule or Checkpoint name ''{0}'' has not been defined y
+
 
 ### Annotators ###
 ANN.keyword.argument.already.provided=Keyword argument already provided: {0}.
