+++ src/main/resources/inspectionDescriptions/SmkSectionMultipleArgsInspection.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+Ony one argument is allowed in subworkflow/shell/wrapper/script/cwl sections.
+</body>
+</html>
