--- src/main/resources/inspectionDescriptions/SmkSubworkflowMultipleArgsInspection.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-Ony one argument is allowed in subworkflow sections
-</body>
-</html>
